#431 fix code 500 if deleted passenger again

See merge request de.pronin/p_dp_airline_1!389

+++ airline-project/src/main/java/app/services/PassengerService.java
@@ -21,7 +21,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class PassengerService {
 
-    @Lazy// FIXME костыль
+    @Lazy
     @Autowired
     private TicketService ticketService;
     @Autowired
@@ -113,6 +113,7 @@ public class PassengerService {
 
     @Transactional
     public void deletePassenger(Long id) {
+        if (getPassenger(id).isEmpty()) throw new EntityNotFoundException("Operation was not finished because Passenger was not found with id = " + id);
         flightSeatService.makeFlightSeatNotSold(ticketService.getFlightSeatIdsByPassengerId(id));
         ticketService.deleteTicketByPassengerId(id);
         bookingService.deleteBookingByPassengerId(id);
